+++ lib/chef/knife/vsphere_vm_delete.rb
@@ -14,53 +14,53 @@ require 'chef/api_client'
 # Delete a virtual machine from vCenter
 class Chef::Knife::VsphereVmDelete < Chef::Knife::BaseVsphereCommand
 
-	banner "knife vsphere vm delete VMNAME"
+  banner "knife vsphere vm delete VMNAME"
 
-	option :purge,
-		:short => "-P",
-		:long => "--purge",
-		:boolean => true,
-		:description => "Destroy corresponding node and client on the Chef Server, in addition to destroying the VM itself."
+  option :purge,
+         :short => "-P",
+         :long => "--purge",
+         :boolean => true,
+         :description => "Destroy corresponding node and client on the Chef Server, in addition to destroying the VM itself."
 
-	get_common_options
+  get_common_options
 
-	# Extracted from Chef::Knife.delete_object, because it has a
-	# confirmation step built in... By specifying the '--purge'
-	# flag (and also explicitly confirming the server destruction!)
-	# the user is already making their intent known. It is not
-	# necessary to make them confirm two more times.
-	def destroy_item(itemClass, name, type_name)
-		object = itemClass.load(name)
-		object.destroy
-		puts "Deleted #{type_name} #{name}"
-	end
+  # Extracted from Chef::Knife.delete_object, because it has a
+  # confirmation step built in... By specifying the '--purge'
+  # flag (and also explicitly confirming the server destruction!)
+  # the user is already making their intent known. It is not
+  # necessary to make them confirm two more times.
+  def destroy_item(itemClass, name, type_name)
+    object = itemClass.load(name)
+    object.destroy
+    puts "Deleted #{type_name} #{name}"
+  end
 
-	def run
-		$stdout.sync = true
+  def run
+    $stdout.sync = true
 
-		vmname = @name_args[0]
+    vmname = @name_args[0]
 
-		if vmname.nil?
-			show_usage
-			fatal_exit("You must specify a virtual machine name")
-		end
+    if vmname.nil?
+      show_usage
+      fatal_exit("You must specify a virtual machine name")
+    end
 
-		vim = get_vim_connection
+    vim = get_vim_connection
 
-		baseFolder = find_folder(get_config(:folder));
+    baseFolder = find_folder(get_config(:folder));
 
-		vm = find_in_folder(baseFolder, RbVmomi::VIM::VirtualMachine, vmname) or
-		fatal_exit("VM #{vmname} not found")
+    vm = find_in_folder(baseFolder, RbVmomi::VIM::VirtualMachine, vmname) or
+        fatal_exit("VM #{vmname} not found")
 
-		vm.PowerOffVM_Task.wait_for_completion unless vm.runtime.powerState == "poweredOff"
-		vm.Destroy_Task
-		puts "Deleted virtual machine #{vmname}"
+    vm.PowerOffVM_Task.wait_for_completion unless vm.runtime.powerState == "poweredOff"
+    vm.Destroy_Task
+    puts "Deleted virtual machine #{vmname}"
 
-		if config[:purge]
-			destroy_item(Chef::Node, vmname, "node")
-			destroy_item(Chef::ApiClient, vmname, "client")
-		else
-			puts "Corresponding node and client for the #{vmname} server were not deleted and remain registered with the Chef Server"
-		end
-	end
+    if config[:purge]
+      destroy_item(Chef::Node, vmname, "node")
+      destroy_item(Chef::ApiClient, vmname, "client")
+    else
+      puts "Corresponding node and client for the #{vmname} server were not deleted and remain registered with the Chef Server"
+    end
+  end
 end
