+++ lib/chef/knife/vsphere_vm_state.rb
@@ -13,9 +13,9 @@ PsOff = 'poweredOff'
 PsSuspended = 'suspended'
 
 PowerStates = {
-  PsOn => 'powered on',
-  PsOff => 'powered off',
-  PsSuspended => 'suspended'
+    PsOn => 'powered on',
+    PsOff => 'powered off',
+    PsSuspended => 'suspended'
 }
 
 # Manage power state of a virtual machine
@@ -26,19 +26,19 @@ class Chef::Knife::VsphereVmState < Chef::Knife::BaseVsphereCommand
   get_common_options
 
   option :state,
-  :short => "-s STATE",
-  :long => "--state STATE",
-  :description => "The power state to transition the VM into; one of on|off|suspended"
+         :short => "-s STATE",
+         :long => "--state STATE",
+         :description => "The power state to transition the VM into; one of on|off|suspended"
 
   option :wait_port,
-    :short => "-w PORT",
-    :long => "--wait-port PORT",
-    :description => "Wait for VM to be accessible on a port"
+         :short => "-w PORT",
+         :long => "--wait-port PORT",
+         :description => "Wait for VM to be accessible on a port"
 
   option :shutdown,
-    :short => "-g",
-    :long => "--shutdown",
-    :description => "Guest OS shutdown"
+         :short => "-g",
+         :long => "--shutdown",
+         :description => "Guest OS shutdown"
 
   def run
 
@@ -56,7 +56,7 @@ class Chef::Knife::VsphereVmState < Chef::Knife::BaseVsphereCommand
     baseFolder = find_folder(get_config(:folder));
 
     vm = find_in_folder(baseFolder, RbVmomi::VIM::VirtualMachine, vmname) or
-      abort "VM #{vmname} not found"
+        abort "VM #{vmname} not found"
 
     state = vm.runtime.powerState
 
@@ -65,45 +65,45 @@ class Chef::Knife::VsphereVmState < Chef::Knife::BaseVsphereCommand
     else
 
       case config[:state]
-      when 'on'
-        if state == PsOn
-          puts "Virtual machine #{vmname} was already powered on"
-        else
-          vm.PowerOnVM_Task.wait_for_completion
-          puts "Powered on virtual machine #{vmname}"
-        end
-      when 'off'
-        if state == PsOff
-          puts "Virtual machine #{vmname} was already powered off"
-        else
-          if get_config(:shutdown)
-            vm.ShutdownGuest
-            print "Waiting for virtual machine #{vmname} to shut down..."
-            until vm.runtime.powerState == PsOff do
-              sleep 2
-              print "."
-            end
-            puts "done"
+        when 'on'
+          if state == PsOn
+            puts "Virtual machine #{vmname} was already powered on"
           else
-            vm.PowerOffVM_Task.wait_for_completion
-            puts "Powered off virtual machine #{vmname}"
+            vm.PowerOnVM_Task.wait_for_completion
+            puts "Powered on virtual machine #{vmname}"
           end
-        end
-      when 'suspend'
-        if state == PowerStates['suspended']
-          puts "Virtual machine #{vmname} was already suspended"
-        else
-          vm.SuspendVM_Task.wait_for_completion
-          puts "Suspended virtual machine #{vmname}"
-        end
-      when 'reset'
-        vm.ResetVM_Task.wait_for_completion
-        puts "Reset virtual machine #{vmname}"
+        when 'off'
+          if state == PsOff
+            puts "Virtual machine #{vmname} was already powered off"
+          else
+            if get_config(:shutdown)
+              vm.ShutdownGuest
+              print "Waiting for virtual machine #{vmname} to shut down..."
+              until vm.runtime.powerState == PsOff do
+                sleep 2
+                print "."
+              end
+              puts "done"
+            else
+              vm.PowerOffVM_Task.wait_for_completion
+              puts "Powered off virtual machine #{vmname}"
+            end
+          end
+        when 'suspend'
+          if state == PowerStates['suspended']
+            puts "Virtual machine #{vmname} was already suspended"
+          else
+            vm.SuspendVM_Task.wait_for_completion
+            puts "Suspended virtual machine #{vmname}"
+          end
+        when 'reset'
+          vm.ResetVM_Task.wait_for_completion
+          puts "Reset virtual machine #{vmname}"
       end
 
       if get_config(:wait_port)
         print "Waiting for port #{get_config(:wait_port)}..."
-        print "." until tcp_test_port_vm(vm,get_config(:wait_port))
+        print "." until tcp_test_port_vm(vm, get_config(:wait_port))
         puts "done"
       end
     end
